+++ src/track/track.service.ts
@@ -3,7 +3,6 @@ import { EErrorMessage } from '../types/messages';
 import { CreateTrackDto } from './dto/createTrack.dto';
 import { UpdateTrackDto } from './dto/updateTrack.dto';
 import { PrismaService } from '../prisma/prisma.service';
-import { Prisma } from '@prisma/client';
 
 @Injectable()
 export class TrackService {
@@ -14,9 +13,9 @@ export class TrackService {
   }
 
   async findOne(id: string) {
-    const currentTrack = this.prisma.track.findUnique({ where: { id } });
+    const currentTrack = await this.prisma.track.findUnique({ where: { id } });
 
-    if (!currentTrack) {
+    if (currentTrack === null) {
       throw new NotFoundException(EErrorMessage.TRACK_NOT_FOUND);
     }
 
@@ -28,33 +27,25 @@ export class TrackService {
   }
 
   async update(id: string, body: UpdateTrackDto) {
-    try {
-      return await this.prisma.track.update({
-        where: { id },
-        data: body,
-      });
-    } catch (error) {
-      if (
-        error instanceof Prisma.PrismaClientKnownRequestError &&
-        error.code === 'P2025'
-      ) {
-        throw new NotFoundException(`Track with id ${id} not found`);
-      }
-      throw error;
+    const currentTrack = await this.prisma.track.findUnique({ where: { id } });
+
+    if (currentTrack === null) {
+      throw new NotFoundException(EErrorMessage.TRACK_NOT_FOUND);
     }
+
+    return this.prisma.track.update({
+      where: { id },
+      data: body,
+    });
   }
 
   async delete(id: string) {
-    try {
-      return await this.prisma.track.delete({ where: { id } });
-    } catch (error) {
-      if (
-        error instanceof Prisma.PrismaClientKnownRequestError &&
-        error.code === 'P2025'
-      ) {
-        throw new NotFoundException(`Track with id ${id} not found`);
-      }
-      throw error;
+    const currentTrack = await this.prisma.track.findUnique({ where: { id } });
+
+    if (currentTrack === null) {
+      throw new NotFoundException(EErrorMessage.TRACK_NOT_FOUND);
     }
+
+    return this.prisma.track.delete({ where: { id } });
   }
 }
